+++ src/nix/hive/flake.nix
@@ -1,13 +1,13 @@
 {
   description = "Internal Colmena expressions";
 
-  outputs = { ... }: {
-    lib.colmenaEval = {
-      rawHive ? null,
-      flakeUri ? null,
-      hermetic ? flakeUri != null,
-    }: import ./eval.nix {
-      inherit rawHive flakeUri hermetic;
+  inputs = {
+    hive.url = "%hive%";
+  };
+
+  outputs = { self, hive }: {
+    colmenaEval = import ./eval.nix {
+      rawFlake = hive;
       colmenaOptions = import ./options.nix;
       colmenaModules = import ./modules.nix;
     };